Fix idempotency for publications
If there are multiple publications of a repository version, then the
rpm_publication module will create another, rather than keeping or
updating one of the existing ones.

This happens due to the find() method on PulpEntity looking for
count==1. It uses a limit of one, so only one result is returned, but
the count still gives the full number of matching items, which in this
case is 2.

This change fixes the issue by failing when we cannot find a unique
entity matching the requested parameters.

Fixes: #84