rpm_sync: support sync_policy
sync_policy is a parameter for RPM syncs which allows switching to a
mirroring mode rather than additive. This can be useful particularly
when using on_demand remotes which may have removed content.

sync_policy was added in pulp_rpm 3.16.0, and replaces the deprecated
mirror parameter. If pulp_rpm is detected to be less than 3.16.0, this
change falls back to use the mirror parameter.